src/requirements.txt
airtable-python-wrapper==0.13.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
docker==4.2.1
fastapi==0.55.1
h11==0.9.0
httptools==0.1.1
idna==2.9
importlib-metadata==1.6.1
mirakuru==2.3.0
more-itertools==8.3.0
mysql-connector-python==8.0.20
mysqlclient==1.4.6
packaging==20.4
pluggy==0.13.1
port-for==0.4
protobuf==3.12.2
psutil==5.7.0
py==1.8.1
pydantic==1.5.1
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.3
websocket-client==0.57.0
websockets==8.1
zipp==3.1.0
